hetiv: Heteroskedasticity and Proxy-Based IV Methods

Tools for identifying structural shocks using heteroskedasticity- and proxy-based instrumental variable (IV) methods in event-study settings. Supports local projection impulse response estimation of Jorda (2005) doi:10.1257/0002828053828518 , generalised weak instrument testing of Lewis and Mertens (2025) doi:10.1093/restud/rdaf103 , and Kalman-filter shock extraction of Burri and Kaufmann (2026) doi:10.1016/j.econlet.2026.113091 .

See also

Author

Maintainer: Daniel Kaufmann info@dankaufmann.com [copyright holder]

Authors:

Marc Burri

Valentin Grob